Minor modifications with a few code cleanup.
Details:
    - Decreasing the timeout value range to 2 bytes.
    - Using seconds instead of miliseconds for timeout indication.
    - Removing unused variable and a couple of variable renaming.

+++ core/src/config.cpp
@@ -16,7 +16,7 @@ Config::Config(const RunMode &mode, const std::string &filePath)
       general_({configYaml_["general"]["fakeUrl"].as<std::string>(),
                 configYaml_["general"]["method"].as<std::string>(),
                 configYaml_["general"]["timeWait"].as<unsigned int>(),
-                configYaml_["general"]["timeout"].as<unsigned int>(),
+                configYaml_["general"]["timeout"].as<unsigned short>(),
                 configYaml_["general"]["repeatWait"].as<unsigned short>()}),
       log_({configYaml_["log"]["logLevel"].as<std::string>(),
             configYaml_["log"]["logFile"].as<std::string>()}),
